Restructured controllers
